[WIP] render tiles with PIL
Mapper sample_shapes.py reads CSV file and produces records of the
format (tilehash, px, py) where tilehash is the QuadTree key of the
tile and px,py are the x,y coordinate of a pixel in that tile.

Reducer draw_tiles.py reads an entire tile worth of points and renders
the image into a memory buffer.